+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Added
 - Tests for Credfeto.Notification.Bot.Twitch.Models to achieve 100% code coverage
 - Added unit test project for Credfeto.Notification.Bot.Server achieving 100% code coverage
+- Tests for Credfeto.Notification.Bot.Twitch.DataTypes assembly - Streamer, Viewer, TwitchUser, and mapper types
 ### Fixed
 ### Changed
 - Dependencies - Updated Credfeto.Enumeration to 1.2.144.1906